Use Arguments also in tests and removed the old impl. Also recoverd tests for errors but implemented them in future module
