+++ rstest_macros/src/lib.rs
@@ -17,7 +17,7 @@ mod utils;
 
 use syn::{parse_macro_input, ItemFn};
 
-use crate::parse::{fixture::FixtureInfo, future::ReplaceFutureAttribute, rstest::RsTestInfo};
+use crate::parse::{fixture::FixtureInfo, rstest::RsTestInfo};
 use parse::ExtendWithFunctionAttrs;
 use quote::ToTokens;
 
@@ -1028,14 +1028,10 @@ pub fn rstest(
     let mut test = parse_macro_input!(input as ItemFn);
     let mut info = parse_macro_input!(args as RsTestInfo);
 
-    let replace_result = ReplaceFutureAttribute::replace(&mut test);
     let extend_result = info.extend_with_function_attrs(&mut test);
 
     let mut errors = error::rstest(&test, &info);
 
-    if let Err(attrs_errors) = replace_result {
-        attrs_errors.to_tokens(&mut errors);
-    }
     if let Err(attrs_errors) = extend_result {
         attrs_errors.to_tokens(&mut errors);
     }
